fix(sjwh): 1:1 redo 6 DMWH features — xzqh, yhwd, swjg, fpbl, skgk, hjfl

git-svn-id: svn://172.16.27.209/gsgh/gsgh/lghjft-backend@1090 63aaa323-bd20-4782-bdcf-f624a29d9816

+++ yudao-module-lghjft/src/main/java/cn/iocoder/yudao/module/lghjft/controller/admin/sjwh/dmwh/hjfl/vo/HjflPageReqVO.java
@@ -21,7 +21,7 @@ public class HjflPageReqVO extends PageParam {
     private String hjflmc;
 
     @Schema(description = "顺序号")
-    private Short sxh;
+    private Long sxh;
 
     @Schema(description = "户籍分类id", example = "27518")
     private Integer hjflid;
