+++ yudao-module-lghjft/src/main/java/cn/iocoder/yudao/module/lghjft/dal/dataobject/sjwh/dmwh/hjfl/HjflDO.java
@@ -40,7 +40,7 @@ public class HjflDO {
     /**
      * 顺序号
      */
-    private Short sxh;
+    private Long sxh;
     /**
      * 户籍分类id
      */
