+++ yudao-module-lghjft/src/main/java/cn/iocoder/yudao/module/lghjft/service/sjwh/dmwh/skgk/SkgkService.java
@@ -15,49 +15,16 @@ import java.util.List;
  */
 public interface SkgkService {
 
-    /**
-     * 创建收款国库
-     *
-     * @param createReqVO 创建信息
-     * @return 编号
-     */
-    Integer createSkgk(@Valid SkgkSaveReqVO createReqVO);
+    Long createSkgk(@Valid SkgkSaveReqVO createReqVO);
 
-    /**
-     * 更新收款国库
-     *
-     * @param updateReqVO 更新信息
-     */
     void updateSkgk(@Valid SkgkSaveReqVO updateReqVO);
 
-    /**
-     * 删除收款国库
-     *
-     * @param id 编号
-     */
-    void deleteSkgk(Integer id);
+    void deleteSkgk(Long id);
 
-    /**
-     * 批量删除收款国库
-     *
-     * @param ids 编号
-     */
-    void deleteSkgkListByIds(List<Integer> ids);
+    void deleteSkgkListByIds(List<Long> ids);
 
-    /**
-     * 获得收款国库
-     *
-     * @param id 编号
-     * @return 收款国库
-     */
-    SkgkDO getSkgk(Integer id);
+    SkgkDO getSkgk(Long id);
 
-    /**
-     * 获得收款国库分页
-     *
-     * @param pageReqVO 分页查询
-     * @return 收款国库分页
-     */
     PageResult<SkgkDO> getSkgkPage(SkgkPageReqVO pageReqVO);
 
 }
